+++ OctopusGame/animatedtile.h
@@ -9,16 +9,16 @@ class GraphicalOctopus;
 
 class AnimatedTile : public Tile {
 public:
-	AnimatedTile(std::vector<Vector2> tilesetPositions, int duration, SDL_Texture* tileset, Vector2 size, Vector2 position);
+	AnimatedTile(std::vector<Vector2> tilesetPositions, float duration, SDL_Texture* tileset, Vector2 size, Vector2 position);
 	void update(float elapsedTime);
 	void draw(GraphicalOctopus &graphics);
 protected:
-	int _amountOfTime = 0;
+	float _amountOfTime = 0;
 	bool _notDone = false;
 private:
 	std::vector<Vector2> _tilesetPositions;
 	int _tileToDraw;
-	int _duration;
+	float _duration;
 };
 
 struct AnimatedTileInfo {
@@ -26,7 +26,7 @@ public:
 	int tilesetFirstGid;
 	int startTileId;
 	std::vector<int> tileIds;
-	int duration;
+	float duration;
 };
 
 #endif // !ANIMATEDTILE_H
